fix refactor bug
