+++ apps/interpolate_sample.py
@@ -14,7 +14,7 @@ from torchvision.utils import save_image
 sys.path.append(os.path.abspath(os.path.join(__file__, '../..')))  # isort:skip  # noqa
 
 from mmgen.apis import set_random_seed # isort:skip  # noqa
-from mmgen.core.evaluation import slerp # isort:skip  # noqa
+from mmgen.engine.evaluation import slerp # isort:skip  # noqa
 from mmgen.models import build_model # isort:skip  # noqa
 
 # yapf: enable
