+++ mmgen/models/architectures/stylegan/generator_discriminator_v2.py
@@ -7,7 +7,7 @@ import torch
 import torch.nn as nn
 from mmcv.runner.checkpoint import _load_checkpoint_with_prefix
 
-from mmgen.core.runners.fp16_utils import auto_fp16
+from mmgen.engine.runners.fp16_utils import auto_fp16
 from mmgen.models.architectures import PixelNorm
 from mmgen.models.architectures.common import get_module_device
 from mmgen.models.builder import build_module
