+++ mmgen/utils/setup_env.py
@@ -16,8 +16,8 @@ def register_all_modules(init_default_scope: bool = True) -> None:
             to https://github.com/open-mmlab/mmengine/blob/main/docs/en/tutorials/registry.md
             Defaults to True.
     """  # noqa
-    import mmgen.core  # noqa: F401,F403
     import mmgen.datasets  # noqa: F401,F403
+    import mmgen.engine  # noqa: F401,F403
     import mmgen.models  # noqa: F401,F403
 
     if init_default_scope:
