+++ tests/test_engine/test_data_structures/test_gen_data_sample.py
@@ -5,7 +5,7 @@ import numpy as np
 import torch
 from mmengine.data import LabelData
 
-from mmgen.core import GenDataSample
+from mmgen.structures import GenDataSample
 
 
 class TestGenDataSample(TestCase):
