+++ tests/test_engine/test_scheduler/test_lr_updater.py
@@ -6,7 +6,7 @@ import torch.nn.functional as F
 import torch.optim as optim
 from mmengine import MessageHub
 
-from mmgen.core.scheduler import LinearLrInterval
+from mmgen.engine.schedulers import LinearLrInterval
 
 
 class ToyModel(torch.nn.Module):
