+++ tests/test_models/test_gans/test_base_conditional_gan.py
@@ -7,7 +7,7 @@ from mmcls.structures import ClsDataSample
 from mmengine.data import LabelData
 from mmengine.testing import assert_allclose
 
-from mmgen.core import GenDataSample
+from mmgen.structures import GenDataSample
 from mmgen.models import BaseConditionalGAN, GANDataPreprocessor
 
 generator = dict(
